Adding backtest_file as a parameter to the function the prints the plots

+++ Statistical-Arbitrage-Bot-Build-in-Crypto/StatBot-PyCharm/Strategy/func_plot_trends.py
@@ -3,13 +3,12 @@ from func_cointegration import calculate_cointegration
 from func_cointegration import calculate_spread
 from func_cointegration import extract_close_prices
 from func_cointegration import calculate_zscore
-from config_strategy_api import backtest_file
 import matplotlib.pyplot as plt
 import pandas as pd
-
+import os.path
 
 # Plot prices and trends
-def plot_trends(sym_1, sym_2, price_data):
+def plot_trends(sym_1, sym_2, price_data, backtest_file):
 
     # Extract prices
     prices_1 = extract_close_prices(price_data[sym_1])
@@ -32,13 +31,14 @@ def plot_trends(sym_1, sym_2, price_data):
     series_2 = df[f"{sym_2}_pct"].astype(float).values
 
     # Save results for backtesting
-    df_2 = pd.DataFrame()
-    df_2[sym_1] = prices_1
-    df_2[sym_2] = prices_2
-    df_2["Spread"] = spread
-    df_2["ZScore"] = zscore
-    df_2.to_csv(backtest_file)
-    print("File for backtesting saved.")
+    if os.path.isfile(backtest_file) == False:
+        df_2 = pd.DataFrame()
+        df_2[sym_1] = prices_1
+        df_2[sym_2] = prices_2
+        df_2["Spread"] = spread
+        df_2["ZScore"] = zscore
+        df_2.to_csv(backtest_file)
+        print("File for backtesting saved.")
 
     # Plot charts
     fig, axs = plt.subplots(3, figsize=(16, 8))
